+++ CHANGES.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed `SyntaxFacts.GetConstantValue` not returning `None` for values outside of the valid range.
 - Fixed the constant folder turning `/` into integer division when both sides are integers.
 - Fixed the constant folder storing the length as an integer token.
+- Fixed `LuaSyntaxNode.GetStructure` throwing an exception because we *do* have structured trivia.
 
 ## v0.2.9-beta.1
 ### Changed